requirements.txt
tk
google-generativeai
pyttsx3